+++ src/main/java/com/app/folioman/pythonbridge/domain/PythonCommand.java
@@ -1,7 +1,6 @@
 package com.app.folioman.pythonbridge.domain;
 
 import java.nio.charset.StandardCharsets;
-import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.List;
 import org.jspecify.annotations.Nullable;
@@ -15,22 +14,6 @@ public record PythonCommand(
         @Nullable byte[] inputData) {
     public PythonCommand {
         arguments = List.copyOf(arguments);
-    }
-
-    public static PythonCommand script(Path scriptPath) {
-        return new PythonCommand(null, scriptPath.toString(), List.of(), null, null, null);
-    }
-
-    public static PythonCommand module(String moduleName) {
-        return new PythonCommand(null, "-m", List.of(moduleName), null, null, null);
-    }
-
-    public static PythonCommand command(String... args) {
-        return new PythonCommand(null, null, List.of(args), null, null, null);
-    }
-
-    public static PythonCommand cli(String executable, String... args) {
-        return new PythonCommand(executable, null, List.of(args), null, null, null);
     }
 
     public PythonCommand withArgument(String arg) {
