+++ src/main/java/com/app/folioman/pythonbridge/domain/PythonTimeoutException.java
@@ -2,7 +2,7 @@ package com.app.folioman.pythonbridge.domain;
 
 public class PythonTimeoutException extends PythonExecutionException {
 
-    public PythonTimeoutException(String message, String command) {
+    PythonTimeoutException(String message, String command) {
         super(message, -1, null, command);
     }
 }
